backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
pytz
tavily-python
google-genai
anyio
httpx
